==> assign1/notebooks/01_preprocessing.ipynb <==
### Imports

import re
import os
import pickle
import pandas as pd

from tqdm import tqdm

from nltk import download
from nltk.corpus import stopwords, gutenberg
from nltk.stem import WordNetLemmatizer

# download('gutenberg')
# download('stopwords')
# download('punkt')

stop_words = set(stopwords.words('english'))
stop_words.remove("very")
stop_words.add("th")
stop_words.add("[")
stop_words.add("]")

tqdm.pandas()
lemmatizer = WordNetLemmatizer()

DATA_DIR = './../data/pickle1/'
OUTPUT_DIR = './../data/processed/'

### Functions

def remove_special_chars(text):
    """
    This function returns a string with filtered special case characters
    """
    return re.sub('[^A-Za-z0-9 ]+', '', text)

def remove_n(text):
    """
    This function returns a string after removing new line char(\n) from a string
    """
    clean1 = re.compile('\n')
    return re.sub(clean1,'', text)

def remove_numbers(text):
    """
    This function returns a string with filtered numbers
    """
    return re.sub('[^A-Za-z ]+', '', text)

### Fetch Data

sentences = []
for file in os.listdir(DATA_DIR):
    if(file.endswith('.pickle')):
#         if(file!='gutenberg_sentences_large.pickle'):
        with open(DATA_DIR + file, 'rb') as f:
            data = pickle.load(f)
        print("sentences: " + str(len(data)))
        sentences += data
        
df = pd.DataFrame(sentences, columns=['sentences'])

len(sentences)

### Convert to lower case

df['sentences'] = df['sentences'].apply(lambda x : x.lower())

### Removing Stopwords

df['sentences'] = df['sentences'].progress_apply(lambda sent : ' '.join([word for word in sent.split() if word not in stop_words]))

### Lemmatization

df['sentences'] = df['sentences'].progress_apply(lambda sent: ' '.join([lemmatizer.lemmatize(word) for word in sent.split()]))

### Special Characters Removal

df['sentences'] = df['sentences'].progress_apply(lambda sent: remove_special_chars(sent))
df['sentences'] = df['sentences'].progress_apply(lambda sent: remove_n(sent))
df['sentences'] = df['sentences'].progress_apply(lambda sent: remove_numbers(sent))

### words with length above threshold removal

df['sentences'] = df['sentences'].progress_apply(lambda x: ' '.join([item for item in x.split() if len(item)<25]))

df = df[df['sentences']!=''].reset_index(drop=True)

### Save Preprocessed Data

df.to_csv(OUTPUT_DIR + 'processed_data_small_lemma.csv',index=False)

# Exploratory Data Analysis

df['sentences'] = df['sentences'].apply(lambda x : x.split())

from collections import Counter
text_counter = Counter()
_ = df['sentences'].apply(text_counter.update)

vocab_old = len(text_counter)
words_old = str(sum(text_counter.values()))

print("No. of unique words present in old processed train data = ",vocab_old)
print("Total no of words : " + words_old)

### Current Vocab size : 60757

##### Removing Redundant words which act as stopwords

text_counter.most_common()

df['sentences'] = df['sentences'].apply(lambda x : x.split())

from collections import Counter
text_counter = Counter()
_ = df['sentences'].apply(text_counter.update)


# print("No. of unique words present in old processed train data = ",vocab_old)
# print("Total no of words : " + words_old)

print("No. of unique words present in processed train data = ",len(text_counter))
print("Total no of words : " + str(sum(text_counter.values())))

remove_words = ['shall', 'ye', 'mr', 'u', 'shalt', 'pdf', 'hast', 'b', 'th', 'oh', 'in', 'de', 'c', 'le']

df['sentences'] = df['sentences'].progress_apply(lambda sent : ' '.join([word for word in sent.split() if word not in remove_words]))

### Save New Preprocessed data

df.to_csv(OUTPUT_DIR + 'processed_data.csv',index=False)

result = ''
for row in df['sentences']:
    row += ' '
    result += row

df = pd.read_csv('./../data/Validation.txt', sep=' ', names=['w1','w2','w3','w4'])
df2 = df[['w3','w4']].rename(columns={'w3':'w1', 'w4':'w2'})
df = pd.concat([df[['w1','w2']],df2], ignore_index=True)
df = df.drop_duplicates()

df3 = df[['w2']].rename(columns={'w2':'w1'})
df = pd.concat([df[['w1']],df2], ignore_index=True)
df = df.drop_duplicates()

df['w1'] = df['w1'].apply(lambda x : x.lower())
# df['w2'] = df['w2'].apply(lambda x : x.lower())

df['w1_present'] = df['w1'].apply(lambda x : x in result)
# df['w2_present'] = df['w2'].apply(lambda x : x in result)

print(df['w1_present'].value_counts())
# print(df['w2_present'].value_counts())
